requirements.txt
duckdb==1.4.3
mcp==1.25.0
pyyaml==6.0.3
requests==2.32.5
weasyprint==67.0
markdown==3.10.1
